+++ guillotina/contrib/audit/interfaces.py
@@ -1,0 +1,5 @@
+from guillotina.async_util import IAsyncUtility
+
+
+class IAuditUtility(IAsyncUtility):
+    pass